change seedmongo to mongo and move jsonmongo into directory
